requirements.txt
google-generativeai==0.8.5
gspread==6.2.1
python-dotenv==1.1.1
python-telegram-bot==22.3
